Add files via upload
This project contains a web application developed using the Python programming language and the Django web framework.

The website provides a platform for users to interact with a specific purpose in mind. The project is designed to create a fast and secure web application by leveraging the powerful features of Django.